@@ -1,7 +1,10 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { App } from './App';
+import { assertNotNullish } from './lib/assert';
 
 window.addEventListener('DOMContentLoaded', () => {
-    ReactDOM.render(<App />, document.getElementById('root'));
+    const container = document.getElementById('root');
+    assertNotNullish(container);
+    const root = createRoot(container);
+    root.render(<App />);
 });
